Add grpc test from #580
